+++ guide/src/getting_started/examples.md
@@ -59,7 +59,7 @@ for integers \\(a\\), \\(b\\) and \\(c\\).
 
 ```rust
 use algebraeon::nzq::Integer;
-use algebraeon::rings::linear::finitely_free_modules::RingToFinitelyFreeModuleStructure;
+use algebraeon::rings::linear::finitely_free_modules::RingToFinitelyFreeModuleSignature;
 use algebraeon::rings::linear::matrix::Matrix;
 use algebraeon::sets::structure::MetaType;
 let m = Matrix::<Integer>::from_rows(vec![vec![3, 4, 1], vec![2, 1, 2], vec![1, 3, -1]]);
