+++ rings/src/polynomial/mod.rs
@@ -2,6 +2,8 @@ mod factoring;
 mod hensel_lifting;
 mod multipoly;
 mod polynomial;
+mod polynomial_ring;
+mod polynomial_semiring;
 mod quotient;
 mod symmetric;
 
@@ -9,5 +11,7 @@ pub use factoring::*;
 pub use hensel_lifting::*;
 pub use multipoly::*;
 pub use polynomial::*;
+pub use polynomial_ring::*;
+pub use polynomial_semiring::*;
 pub use quotient::*;
 pub use symmetric::*;
