+++ rings/src/rings/algebraic_number_fields/ring_of_integers.rs
@@ -1,7 +1,7 @@
 use super::number_field::AlgebraicNumberFieldStructure;
 use crate::{
     linear::{
-        finitely_free_modules::{FinitelyFreeModuleStructure, RingToFinitelyFreeModuleStructure},
+        finitely_free_modules::{FinitelyFreeModuleStructure, RingToFinitelyFreeModuleSignature},
         matrix::Matrix,
     },
     polynomial::Polynomial,
